requirements.txt
matplotlib==3.10.3
numpy==2.3.0
onnx==1.18.0
onnxruntime_gpu==1.22.0
onnxscript==0.3.0
pandas==2.3.0
peft==0.15.1
Pillow==11.2.1
rustworkx==0.16.0
scikit_learn==1.7.0
scipy==1.15.3
timm==1.0.15
torch==2.4.1
torchinfo==1.8.0
torchvision==0.19.1
